moved CoordinateSystem to own file
